Abstract Code More Towards MVC for Further Customization Capability
