+++ settings.toml
@@ -4,5 +4,5 @@ CIRCUITPY_WIFI_PASSWORD = "<PASSWORD>"
 511_API_KEY = "<API_KEY>"
 511_TRANSIT_AGENCY = "SF"
 511_TRANSIT_DIRECTION = "OB"
-511_TRANSIT_ROUTE_CODES = "K,N"
+511_TRANSIT_ROUTE_CODES = "J,K,M,N"
 511_TRANSIT_STOP_CODE = "16995"
